All navigation views should be written in navigation.js (#45)
